+++ packages/experimental/src/parties/test-party-processor.ts
@@ -13,13 +13,25 @@ import { PartyProcessor } from './party-processor';
  */
 export class TestPartyProcessor extends PartyProcessor {
   async _processMessage (message: IHaloStream): Promise<void> {
-    const { data: { genesis } } = message;
+    const { data: { genesis, admit } } = message;
 
     //
     // Party genesis.
     //
     if (genesis) {
       const { partyKey, feedKey } = genesis;
+      assert(partyKey);
+      assert(feedKey);
+      assert(Buffer.compare(partyKey, this._partyKey) === 0);
+      this._feedKeys.add(feedKey);
+      return;
+    }
+
+    //
+    // Admit feed.
+    //
+    if (admit) {
+      const { partyKey, feedKey } = admit;
       assert(partyKey);
       assert(feedKey);
       assert(Buffer.compare(partyKey, this._partyKey) === 0);