+++ packages/experimental/src/proto/gen/testing.js
@@ -769,6 +769,7 @@ $root.dxos = (function() {
                  * Properties of a PartyAdmit.
                  * @memberof dxos.echo.testing
                  * @interface IPartyAdmit
+                 * @property {Uint8Array|null} [partyKey] PartyAdmit partyKey
                  * @property {Uint8Array|null} [feedKey] PartyAdmit feedKey
                  */
 
@@ -786,6 +787,14 @@ $root.dxos = (function() {
                             if (properties[keys[i]] != null)
                                 this[keys[i]] = properties[keys[i]];
                 }
+
+                /**
+                 * PartyAdmit partyKey.
+                 * @member {Uint8Array} partyKey
+                 * @memberof dxos.echo.testing.PartyAdmit
+                 * @instance
+                 */
+                PartyAdmit.prototype.partyKey = $util.newBuffer([]);
 
                 /**
                  * PartyAdmit feedKey.
@@ -819,8 +828,10 @@ $root.dxos = (function() {
                 PartyAdmit.encode = function encode(message, writer) {
                     if (!writer)
                         writer = $Writer.create();
+                    if (message.partyKey != null && Object.hasOwnProperty.call(message, "partyKey"))
+                        writer.uint32(/* id 1, wireType 2 =*/10).bytes(message.partyKey);
                     if (message.feedKey != null && Object.hasOwnProperty.call(message, "feedKey"))
-                        writer.uint32(/* id 1, wireType 2 =*/10).bytes(message.feedKey);
+                        writer.uint32(/* id 2, wireType 2 =*/18).bytes(message.feedKey);
                     return writer;
                 };
 
@@ -856,6 +867,9 @@ $root.dxos = (function() {
                         var tag = reader.uint32();
                         switch (tag >>> 3) {
                         case 1:
+                            message.partyKey = reader.bytes();
+                            break;
+                        case 2:
                             message.feedKey = reader.bytes();
                             break;
                         default:
@@ -893,6 +907,9 @@ $root.dxos = (function() {
                 PartyAdmit.verify = function verify(message) {
                     if (typeof message !== "object" || message === null)
                         return "object expected";
+                    if (message.partyKey != null && message.hasOwnProperty("partyKey"))
+                        if (!(message.partyKey && typeof message.partyKey.length === "number" || $util.isString(message.partyKey)))
+                            return "partyKey: buffer expected";
                     if (message.feedKey != null && message.hasOwnProperty("feedKey"))
                         if (!(message.feedKey && typeof message.feedKey.length === "number" || $util.isString(message.feedKey)))
                             return "feedKey: buffer expected";
@@ -911,6 +928,11 @@ $root.dxos = (function() {
                     if (object instanceof $root.dxos.echo.testing.PartyAdmit)
                         return object;
                     var message = new $root.dxos.echo.testing.PartyAdmit();
+                    if (object.partyKey != null)
+                        if (typeof object.partyKey === "string")
+                            $util.base64.decode(object.partyKey, message.partyKey = $util.newBuffer($util.base64.length(object.partyKey)), 0);
+                        else if (object.partyKey.length)
+                            message.partyKey = object.partyKey;
                     if (object.feedKey != null)
                         if (typeof object.feedKey === "string")
                             $util.base64.decode(object.feedKey, message.feedKey = $util.newBuffer($util.base64.length(object.feedKey)), 0);
@@ -932,7 +954,14 @@ $root.dxos = (function() {
                     if (!options)
                         options = {};
                     var object = {};
-                    if (options.defaults)
+                    if (options.defaults) {
+                        if (options.bytes === String)
+                            object.partyKey = "";
+                        else {
+                            object.partyKey = [];
+                            if (options.bytes !== Array)
+                                object.partyKey = $util.newBuffer(object.partyKey);
+                        }
                         if (options.bytes === String)
                             object.feedKey = "";
                         else {
@@ -940,6 +969,9 @@ $root.dxos = (function() {
                             if (options.bytes !== Array)
                                 object.feedKey = $util.newBuffer(object.feedKey);
                         }
+                    }
+                    if (message.partyKey != null && message.hasOwnProperty("partyKey"))
+                        object.partyKey = options.bytes === String ? $util.base64.encode(message.partyKey, 0, message.partyKey.length) : options.bytes === Array ? Array.prototype.slice.call(message.partyKey) : message.partyKey;
                     if (message.feedKey != null && message.hasOwnProperty("feedKey"))
                         object.feedKey = options.bytes === String ? $util.base64.encode(message.feedKey, 0, message.feedKey.length) : options.bytes === Array ? Array.prototype.slice.call(message.feedKey) : message.feedKey;
                     return object;
@@ -965,6 +997,7 @@ $root.dxos = (function() {
                  * Properties of a PartyEject.
                  * @memberof dxos.echo.testing
                  * @interface IPartyEject
+                 * @property {Uint8Array|null} [partyKey] PartyEject partyKey
                  * @property {Uint8Array|null} [feedKey] PartyEject feedKey
                  */
 
@@ -982,6 +1015,14 @@ $root.dxos = (function() {
                             if (properties[keys[i]] != null)
                                 this[keys[i]] = properties[keys[i]];
                 }
+
+                /**
+                 * PartyEject partyKey.
+                 * @member {Uint8Array} partyKey
+                 * @memberof dxos.echo.testing.PartyEject
+                 * @instance
+                 */
+                PartyEject.prototype.partyKey = $util.newBuffer([]);
 
                 /**
                  * PartyEject feedKey.
@@ -1015,8 +1056,10 @@ $root.dxos = (function() {
                 PartyEject.encode = function encode(message, writer) {
                     if (!writer)
                         writer = $Writer.create();
+                    if (message.partyKey != null && Object.hasOwnProperty.call(message, "partyKey"))
+                        writer.uint32(/* id 1, wireType 2 =*/10).bytes(message.partyKey);
                     if (message.feedKey != null && Object.hasOwnProperty.call(message, "feedKey"))
-                        writer.uint32(/* id 1, wireType 2 =*/10).bytes(message.feedKey);
+                        writer.uint32(/* id 2, wireType 2 =*/18).bytes(message.feedKey);
                     return writer;
                 };
 
@@ -1052,6 +1095,9 @@ $root.dxos = (function() {
                         var tag = reader.uint32();
                         switch (tag >>> 3) {
                         case 1:
+                            message.partyKey = reader.bytes();
+                            break;
+                        case 2:
                             message.feedKey = reader.bytes();
                             break;
                         default:
@@ -1089,6 +1135,9 @@ $root.dxos = (function() {
                 PartyEject.verify = function verify(message) {
                     if (typeof message !== "object" || message === null)
                         return "object expected";
+                    if (message.partyKey != null && message.hasOwnProperty("partyKey"))
+                        if (!(message.partyKey && typeof message.partyKey.length === "number" || $util.isString(message.partyKey)))
+                            return "partyKey: buffer expected";
                     if (message.feedKey != null && message.hasOwnProperty("feedKey"))
                         if (!(message.feedKey && typeof message.feedKey.length === "number" || $util.isString(message.feedKey)))
                             return "feedKey: buffer expected";
@@ -1107,6 +1156,11 @@ $root.dxos = (function() {
                     if (object instanceof $root.dxos.echo.testing.PartyEject)
                         return object;
                     var message = new $root.dxos.echo.testing.PartyEject();
+                    if (object.partyKey != null)
+                        if (typeof object.partyKey === "string")
+                            $util.base64.decode(object.partyKey, message.partyKey = $util.newBuffer($util.base64.length(object.partyKey)), 0);
+                        else if (object.partyKey.length)
+                            message.partyKey = object.partyKey;
                     if (object.feedKey != null)
                         if (typeof object.feedKey === "string")
                             $util.base64.decode(object.feedKey, message.feedKey = $util.newBuffer($util.base64.length(object.feedKey)), 0);
@@ -1128,7 +1182,14 @@ $root.dxos = (function() {
                     if (!options)
                         options = {};
                     var object = {};
-                    if (options.defaults)
+                    if (options.defaults) {
+                        if (options.bytes === String)
+                            object.partyKey = "";
+                        else {
+                            object.partyKey = [];
+                            if (options.bytes !== Array)
+                                object.partyKey = $util.newBuffer(object.partyKey);
+                        }
                         if (options.bytes === String)
                             object.feedKey = "";
                         else {
@@ -1136,6 +1197,9 @@ $root.dxos = (function() {
                             if (options.bytes !== Array)
                                 object.feedKey = $util.newBuffer(object.feedKey);
                         }
+                    }
+                    if (message.partyKey != null && message.hasOwnProperty("partyKey"))
+                        object.partyKey = options.bytes === String ? $util.base64.encode(message.partyKey, 0, message.partyKey.length) : options.bytes === Array ? Array.prototype.slice.call(message.partyKey) : message.partyKey;
                     if (message.feedKey != null && message.hasOwnProperty("feedKey"))
                         object.feedKey = options.bytes === String ? $util.base64.encode(message.feedKey, 0, message.feedKey.length) : options.bytes === Array ? Array.prototype.slice.call(message.feedKey) : message.feedKey;
                     return object;