+++ packages/experimental/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [08-17-19] PartyManager real-time update and coldstart.
 - [08-17-20] Pipeline with feed-store-iterator (using party-processor).
 - [08-17-20] Pipeline with basic party-procesor.
 - [08-17-20] Working spacetime message order.
@@ -25,10 +26,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Backlog
 
+- Item models (properties and data).
+- Party properties (via special item).
 - Ensure streams are closed when objects are destroyed (on purpose or on error); error handling. Asserts vs Errors?
-- Initial state to bootstrap parties (from FeedStore metadata).
-- Item models (system and user).
-- Party metadata processing (via custom item).
 - Reactive components (Database, Party, Item, Model) with event propagation.
 - Event handlers: global state to warn of leaks when system shuts down (show graph).